+++ pyroglyph/exceptions.py
@@ -1,6 +1,19 @@
 # -*- coding: utf-8 -*-
-__all__ = ('__version__', 'PyroglyphException')
+__all__ = ('__version__', 'PyroglyphError', 'AlreadyStartedError',
+           'AlreadyTerminatedError', 'NotStartedError')
 
 
-class PyroglyphException(Exception):
-    """Base class used by all Pyroglyph exceptions."""
+class PyroglyphError(Exception):
+    """Base class used by all Pyroglyph errors."""
+
+
+class NotStartedError(PyroglyphError):
+    """The render loop has not yet been started."""
+
+
+class AlreadyStartedError(PyroglyphError):
+    """The render loop has already been started."""
+
+
+class AlreadyTerminatedError(PyroglyphError):
+    """The render loop has already been terminated."""